refactor(cubestore): extract EXPLAIN ANALYZE DETAILED into its own module
Move explain_detailed (orchestration) and render_query_trace (pure QueryTrace ->
DataFrame rendering) out of the oversized sql/mod.rs into sql/explain_detailed.rs.
render stays out of trace.rs to keep that a DataFrame-free serde leaf.
